requirements.txt
# W17D4 Demo Code Requirements
# Core dependencies for CLIP, BLIP, and Whisper demos

# Core Requirements
transformers>=4.36.0
torch>=2.1.0
pillow>=10.0.0
numpy>=1.24.0

# Optional (for Whisper demo)
librosa>=0.10.0
soundfile>=0.12.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
